Reworked the loading of objects from the Database. The new system now uses special DB_Objects that handle the Loading/Saving/Delete operations. The objects used in-program are now implementing ISaveable and IConnectable. The first handles conversion between the DB_Objects and the normal Objects, and the second handles the connectability between objects, e.g. adding of Entries to an Account.
This was necessary, because loading data in the previous model led to circular references that resulted in an infinite loop.

Additionally, started to implement INotifyChanged on objects, so that we can see those changes in the Grid.

Finally, added a DataHelper class that keeps track of all in-program objects. This is necessary to correctly "connect" objects to each other.